Use package resource defaults rather than setting provider and source
This commit refactors package.pp to use package defaults for provider
and source rather than setting the corresponding attributes in the
package resource declaration. This avoids ever setting the provider and
source attributes to 'undef' in the package declaration.

This shouldn't have any visible effect from a user prospective,
however it will help the users of rspec-puppet working around the
issue described in rodjek/rspec-puppet#256