break up restart_on_change into discrete params and change default
restart_on_change is now a top-level control for the other three discrete
restart parameters (restart_config_change, restart_package_change, and
restart_plugin_change). If restart_on_change is explicity set in
addition to a more specific paramter, the more specific parameter
"wins".

In addition, the top-level parameter default has been changed to `false`
in order to reduce unexpected restarts for users that expect
highly-available behavior out-of-the-box (this can still be changed to
`true` to preserve previous behavior).