Add deprecation log feature
Currently it's not possible to enable the deprecation log, because the
logging template is missing the necessary appender.

This commit will add two new parameters to enable deprecation logging
and to specify a logging level.
The boolean parameter `deprecation_logging` enables deprecation
loggging if set to true.
The string parameter `deprecation_logging_level` specifies the desired
logging level for the deprecation log.