manifests: properly reference platform datadir defaults
The old module referenced the DATA_DIR value straight from ::params, which meant
that it was never changed, which while accurate, means it should not be
referenced as a parameter to the init class file. This sets DATA_DIR to a static
value per-platform, which is the right way of doing it.