(#1203) Fix lib directory permissions without changing the mode of files in the directory
This module changes the file mode of files in /usr/share/elasticsearch/lib everytime
Elasticsearch is upgraded with Apt. This is unnecessary and undesirable.

From this commit https://github.com/voxpupuli/puppet-elasticsearch/commit/854d9022b45e49c00f8b6f7e709a1f1b4372a005,
the intention was to fix the permission on the directory, but it is now
changing the file mode of all the files in the directory as well.

From tests it is alright to set the file mode on the directory to
0644 as Puppet does the right thing setting the directory mode to 0755,
irrespective of the umask value, while setting the file mode to 0644 as
expected.